Fixed bug with build.py due to the src/ structure change.
Switched to using poetry in the makefile for consistency and added more commands.

Updated poetry.lock.

Removed incomplete tasks for pull request.

Changed tests fixtures so the tests only have access to what they need (some files like test_textures.py do have access to the points which they don't need, however, this can't be changed).
